requirements.txt
streamlit>=1.28.0
pymongo>=4.5.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-mongodb>=0.1.0
langchain-openai>=0.0.8
openai>=1.0.0
pypdf>=3.15.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
certifi>=2023.11.17